requirements.txt
screeninfo==0.6.7
Pillow==8.3.2
